Split GridScrollbar into separate component
